% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_ap.R
\name{theme_ap}
\alias{theme_ap}
\title{ggplot AP theme}
\usage{
theme_ap()
}
\description{
This function is a theme function for use with ggplot2.
It tries to have a very tight composition, suitable for email or chat message
}
